Added files and changes required for deployment to heroku

+++ Emailer/forms.py
@@ -4,7 +4,7 @@ from django import forms
 class EmailerForm(forms.Form):
     subject = forms.CharField(max_length=100)
     message = forms.CharField(widget=forms.Textarea)
-    seconds = forms.IntegerField(min_value=1)
+    seconds = forms.FloatField(min_value=1.0)
     to_whom = forms.EmailField()
 
     subject.widget.attrs.update({'class': 'form-control'})
